- 상품 구매당시 가격, 수량 추가

+++ store/src/main/java/com/min/store/order/domain/OrderItem.java
@@ -18,7 +18,7 @@ public class OrderItem {
     private ProductId product;
 
     @Column
-    private int orderPrice;
+    private Long orderPrice;
 
     @Column
     private int quantity;
